requirements.txt
APScheduler==3.7.0	
aiohttp==3.7.4
beautifulsoup4==4.9.3
discord.py==2.1.1
humanfriendly==9.1
matplotlib==3.3.1
psycopg2==2.8.5
python-dotenv==0.15.0
psutil==5.8.0